Move QueryProcessor into new name and independent file
